No need for an answer of this size

+++ builtin.h
@@ -3,19 +3,19 @@
 
 #include <stdint.h>
 
-unsigned builtin_8(uint8_t c) {
+uint8_t builtin_8(uint8_t c) {
   return __builtin_popcount(c);
 }
 
-unsigned builtin_16(uint16_t s) {
+uint8_t builtin_16(uint16_t s) {
   return __builtin_popcount(s);
 }
 
-unsigned builtin_32(uint32_t l) {
+uint8_t builtin_32(uint32_t l) {
   return __builtin_popcountl(l);
 }
 
-unsigned builtin_64(uint64_t ll) {
+uint8_t builtin_64(uint64_t ll) {
   return __builtin_popcountll(ll);
 }
